Fix core→ui cross-boundary deps via dependency injection
action-context.js: remove static import of ui/store.js; accept openStore
as an optional parameter (no-op default for headless/test contexts).
main.js passes the real openDarknetsStore function.

cheats.js: remove static import of ui/save-load.js; accept saveGame as
an optional { saveGame } parameter. The relayout and restore cheats
(which require DOM or graph.js) are moved out of cheats.js entirely.

console.js: handle "cheat relayout" and "cheat restore" directly
(both are browser-specific). For "cheat snapshot", import save-load.js
and pass saveGame to handleCheatCommand.
